(fix|COS-3350): Add service start date validation, discard changes if date is not valid and display error message in toast
