+++ packages/apps/frontera/src/routes/organization/src/components/Tabs/panels/AccountPanel/Contract/ContractBillingDetailsModal/Services/components/ServiceCard.tsx
@@ -174,6 +174,7 @@ export const ServiceCard: React.FC<ServiceCardProps> = observer(
               contractStatus={contractStatus}
               billingEnabled={billingEnabled}
               allowIndividualRestore={allowIndividualRestore}
+              allServices={data}
             />
           ))}
         </CardContent>
