+++ packages/apps/frontera/src/routes/organization/src/components/Tabs/panels/AccountPanel/Contract/ContractBillingDetailsModal/Services/components/ServiceItem/index.ts
@@ -1,0 +1,1 @@
+export { ServiceItem } from './ServiceItem.tsx';
